+++ src/cvanet/alexnet/jepa/factory.py
@@ -1,5 +1,6 @@
 from typing import Tuple
 from .model import JEPA, Config
+from cvanet.alexnet.backbone.model import init_weights_he
 from cvanet.alexnet.backbone.factory import alexnet_backbone
 
 
@@ -8,6 +9,7 @@ def jepa(config: Config=None, **kwargs) -> Tuple[JEPA, Config]:
         config = Config()
     config.__dict__.update(kwargs)
     backbone, _ = alexnet_backbone(config.backbone)
+    backbone.apply(init_weights_he)
     model = JEPA(
         backbone=backbone, latent_dim=config.latent_dim, ema_tau_min=config.ema_tau_min, ema_tau_max=config.ema_tau_max,
         ema_total_steps=config.ema_total_steps)